% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cycle.R
\name{cycle}
\alias{cycle}
\title{Predict Cyclizability}
\usage{
cycle(sequences, smooth)
}
\arguments{
\item{sequences}{A list or vector of sequences}

\item{smooth}{Whether to predict smoothed C0 (DNAcycP2) or original C0 (DNAcycP)}
}
\value{
A list of (1) predictions on a normalized scale, (2) predictions on an unnormalized scale
}
\description{
This predicts cyclizability for a set of sequences.
}
\examples{
# Example usage of cycle
cycle(sequences, smooth=TRUE) # where sequences is a list/vector of sequences
}
