% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cycle.R
\name{cycle_fasta}
\alias{cycle_fasta}
\title{Predict Cyclizability}
\usage{
cycle_fasta(file_path)
}
\arguments{
\item{input_file}{.fasta input file path}
}
\value{
A list of predictions for each ID in the .fasta file. Each list item has
the following columns: position, c_score_norm (predictions on a normalized scale),
and c_score_unnorm (predictions on an unnormalized scale)
}
\description{
This predicts cyclizability for all subsequences of length 50bp from a .fasta input file.
}
