% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cycle.R
\name{cycle_fasta}
\alias{cycle_fasta}
\title{Predict Cyclizability}
\usage{
cycle_fasta(file_path, smooth, n_cores = 1, chunk_length = 1e+05)
}
\arguments{
\item{smooth}{Whether to predict smoothed C0 (DNAcycP2) or original C0 (DNAcycP)}

\item{n_cores}{Number of cores to use for parallel processing (default=1)}

\item{chunk_length}{Length of sequence that each core will predict on at a given time.
(default=100000)}

\item{input_file}{.fasta input file path}
}
\value{
A list of predictions for each ID in the .fasta file.

Each list item has the following columns: position, c_score_norm (predictions
on a normalized scale), and c_score_unnorm (predictions on an unnormalized
scale).

Each list item is named "cycle_{id}" corresponding to the fasta id
}
\description{
This predicts cyclizability for all subsequences of length 50bp from a .fasta input file.
}
\examples{
# Example usage of cycle_fasta
cycle_fasta("path/to/fasta/file.fasta",smooth=TRUE, n_cores=2, chunk_length=50000)
}
